Implemented Create Doubly Linked List Method
This Method first checks if the DLL is empty. If it is, it creates the first node where the next and previous pointer are set to NULL. This is because it is the first  and only item in the list. The first and only item neither points to anywhere nor is pointed from anywhere.  If the DLL is not empty. The method simply adds the new node to the end of the list.